# PERF: 2x ~ 30x faster dirty tracking