@shioyama 2019-04-21
Performance boost is absolutely fantastic, but wondering about this:

> Currently, although using both dirty tracking (ivar backed and
attributes backed) on one model is not supported (doesn't fully work at
least)

Is this something that would be welcomed as a PR, if it had no impact on performance (i.e. if it could maintain the perf here)? I've hacked around this area in [Mobility](https://github.com/shioyama/mobility) for a while and every change to ActiveRecord requires a new hack.

I personally think it's really a shame that the combination of ivar-backed dirty attributes and attribute-backed attributes is not supported on AR models. It means that the kind of thing users want from my gem is very hard to support without ugly monkey-patching.