@kamipo 2019-04-22
I've read the code https://github.com/shioyama/mobility/blob/master/lib/mobility/plugins/active_record/dirty.rb and https://github.com/shioyama/mobility/blob/master/lib/mobility/plugins/active_model/dirty.rb,
it already deeply hackes/depends on internal API and internal state (e.g. `model.send(:attributes_changed_by_setter).except!(locale_accessor)` should `model.clear_attribute_changes(*locale_accessor)`, hacking force tracking both (ivar/attributes) previous changes, etc.).

I'm not prefer to guarantee such dirty hacks.