@shioyama 2019-04-22
To be clear, #31394 *would have* allowed this by exposing a module builder `ActiveModel::AttributeMethodsBuilder` whose instances could have been used alongside instances of `ActiveRecord::AttributeMethodsBuilder` in the model's ancestors. It also encapsulated internals of the dirty modules into the module builder so they were not exposed in the model class itself.