@shioyama 2019-04-22
> e.g. model.send(:attributes_changed_by_setter).except!(locale_accessor) should be model.clear_attribute_changes(*locale_accessor), hacking force tracking both (ivar/attributes) previous changes, etc.).

I'm pretty sure I tried this and it did not work (at the time anyway). I have looked quite carefully at the AR code and tried hard to make the hack as minimal as possible. It is *very* tricky...