@jrafanie 2019-05-21
Note, this PR indirectly "fixed" the predicate methods `will_save_change_to_attribute?` and `saved_change_to_attribute?` (and probably others) returning nil instead of false.  It's a small change that eliminates the need for me to use `!!` in my code.  Thanks @kamipo !!