@welearnednothing 2019-05-27
I'm trying to use Neo4j with 6.0.0.rc1 but when I try to do a migration (`rails neo4j:migrate`) I get an error `NoMethodError: undefined method `force_change' for #<ActiveModel::NullMutationTracker:0x00007fbed543e338>`. It [looks like I'm not alone](https://stackoverflow.com/questions/55886060/rails-6-and-neo4j-rb).

This commit removed the methods `forget_change`, `original_value`, and `force_change` from `ActiveModel::NullMutationTracker`, making it no longer API compatible with `ActiveModel::AttributeMutationTracker` or the new `ActiveModel::ForcedMutationTracker`. These three methods were no-ops, so it may have seemed as though they were no longer necessary, but now use cases that relied on NullMutationTracker and called those methods will fail.

Was there a motivation for deleting these methods, or were they seen as not needed since they appeared to do nothing? I'm happy to put together a pull request putting them back in place if this was an oversight or push up an example repo of the failure if it's helpful.