@kamipo 2019-05-28
The motivation that removed `forget_change` and `force_change` in the NullMutationTracker is to disallow any mutation to the `mutations_before_last_save` mutation tracker.

If Neo4j will mutate `mutations_before_last_save` or use `NullMutationTracker` internal class directly, it is already something wrong.

Seems it should be addressed in the Neo4j side to me.