@kamipo 2019-05-28
Perhaps reverting the commit https://github.com/neo4jrb/neo4j/commit/3b1d1bf26d9464c126834ddc2267910d3a5b73d9 will avoid the NoMethodError.