@welearnednothing 2019-06-04
@kamipo Thank you so much for the follow up and pointing me to that Neo4j commit - I'm following up with the author of that commit. Are you saying `NullMutationTracker` shouldn't be used directly outside of Rails?