@kamipo 2019-06-04
> Are you saying `NullMutationTracker` shouldn't be used directly outside of Rails?

Yes, that is an internal class `# :nodoc:`ed explicitly.