@welearnednothing 2019-06-04
Got it, I'll work on getting this fixed in the Neo4j gem. Thanks for your help, @kamipo!