@Joshfindit 2019-08-21
Which module needs to be required in order to support the `force_change` method?
Using Sinatra.